+++ lib/gorge/models/player.rb
@@ -17,25 +17,25 @@ module Gorge
       total_stats = player_rounds_dataset.
         select {
           [
-            (sum(hits).cast(:float) / (sum(hits) + sum(misses))).as(accuracy),
-            (sum(kills).cast(:float) / sum(deaths)).as(kdr)
+            (sum(hits).cast(:float) / Sequel.case({ { (sum(hits) + sum(misses)) => 0 } => 1 }, sum(hits) + sum(misses))).as(accuracy),
+            (sum(kills).cast(:float) / Sequel.case({ { sum(:deaths) => 0 } => 1 }, sum(:deaths))).as(kdr),
           ]
       }.first
 
       alien_stats = alien_rounds_dataset.
         select {
           [
-            (sum(hits).cast(:float) / (sum(hits) + sum(misses))).as(accuracy),
-            (sum(kills).cast(:float) / sum(deaths)).as(kdr)
+            (sum(hits).cast(:float) / Sequel.case({ { (sum(hits) + sum(misses)) => 0 } => 1 }, sum(hits) + sum(misses))).as(accuracy),
+            (sum(kills).cast(:float) / Sequel.case({ { sum(:deaths) => 0 } => 1 }, sum(:deaths))).as(kdr),
           ]
       }.first
 
       marine_stats = marine_rounds_dataset.
         select {
           [
-            (sum(hits).cast(:float) / (sum(hits) + sum(misses))).as(accuracy),
-            ((sum(hits).cast(:float) - sum(onos_hits)) / (sum(hits) - sum(onos_hits) + sum(misses))).as(accuracy_no_onos),
-            (sum(kills).cast(:float) / sum(deaths)).as(kdr)
+            (sum(hits).cast(:float) / Sequel.case({ { (sum(hits) + sum(misses)) => 0 } => 1 }, sum(hits) + sum(misses))).as(accuracy),
+            (sum(kills).cast(:float) / Sequel.case({ { sum(:deaths) => 0 } => 1 }, sum(:deaths))).as(kdr),
+            ((sum(hits).cast(:float) - sum(onos_hits)) / Sequel.case({ { (sum(hits) - sum(onos_hits) + sum(misses)) => 0 } => 1 }, sum(hits) - sum(onos_hits) + sum(misses))).as(accuracy_no_onos),
           ]
       }.first