+++ CHANGELOG.md
@@ -11,6 +11,7 @@ glance - what to expact from upgrading to a new version.
 - Importer: Store round lengths.
 - Importer: Store map names.
 - Importer: Store marine and alien starting locations.
+- Importer: Store time spent per player for each class (lifeform, weapon, ...)
 
 ### Changed